+++ android/src/oldarch/java/io/sentry/react/RNSentryModule.java
@@ -24,6 +24,21 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
+    public void addListener(String eventType) {
+        this.impl.addListener(eventType);
+    }
+
+    @ReactMethod
+    public void removeListeners(double id) {
+        this.impl.removeListeners(id);
+    }
+
+    @ReactMethod
+    public void initNativeReactNavigationNewFrameTracking(Promise promise) {
+        this.impl.initNativeReactNavigationNewFrameTracking(promise);
+    }
+
+    @ReactMethod
     public void initNativeSdk(final ReadableMap rnOptions, Promise promise) {
         this.impl.initNativeSdk(rnOptions, promise);
     }