+++ ios/RNSentryEvents.m
@@ -1,0 +1,3 @@
+#import "RNSentryEvents.h"
+
+NSString *const RNSentryNewFrameEvent = @"rn_sentry_new_frame";